fix: implement centralized authentication state management
- Replace passthrough AuthProvider with proper React context
- Move authentication logic from useAuth hook to AuthProvider
- Ensure OAuth callback processing happens at app level
- Fix dashboard not auto-refreshing after successful authentication
- Clean up OAuth parameters from URL after processing
- Maintain backward compatibility with existing useAuth hook

This resolves the issue where dashboard components wouldn't
automatically refresh after authentication because each
component had its own isolated auth state.

Fixes: #2
